app/requirements.txt
streamlit
pandas
scikit-learn
surprise
IMDbPY
matplotlib
